Conflicts:
	Democrazik/src/main/java/fr/democrazik/DemocrazikApplication.java
	Democrazik/src/main/java/fr/democrazik/dao/MorceauRepository.java
	Democrazik/src/main/java/fr/democrazik/web/MorceauRest.java

+++ Democrazik/src/main/java/fr/democrazik/dao/ArtisteRepository.java
@@ -1,6 +1,7 @@
 package fr.democrazik.dao;
 
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -14,7 +15,9 @@ public interface ArtisteRepository extends JpaRepository<Artiste, Long>{
 
 	public List<Artiste> findArtisteByNom(String nom);
 	
+
 	@Query(value = "select artiste.id from artiste left join morceau on artiste.id = morceau.artiste_id where artiste_id IS NULL", 
 	       nativeQuery = true)
-	public List<Long> idOfUselessArtists();
+	public List<BigInteger> idOfUselessArtists();
 }
+
